implemented stopwatch

+++ app/src/main/java/com/better_computer/habitaid/data/AbstractModel.java
@@ -91,7 +91,7 @@ public class AbstractModel {
             return null;
         }else{
             try{
-                return (int)Double.parseDouble(value);
+                return Integer.parseInt(value);
             }catch (Exception e){
                 return null;
             }
